Force user to remove the stick before proceeding with the boot process.
This is especially important for those Qubes users who have a separate USB VM,
and who should avoid exposing the stick to this not-so-trusted domain in order
to avoid a potential attack where the adversary compromised the USB VM
and later is able to make a copy of the Anti Evil Maid stick. While this alone
should not allow the attacker to bypass Anti Evil Maid protection (if the user
also uses SRK password), it might be fatal in case the adversary can also sniff
the user's keystrokes (e.g. via electromagentic leaks, hidden cameras, etc).

Users of traditional OSes that do not support untrusted USB domains are not
any more secure. It's just that those users have no other way of handling the
untrusted USB devices, and so such precutions do not make any sense for them.

For those that decided to install Anti Evil Maid on the HDD, there is an
option: rd.antievilmaid.dontforcestickremoval, that can be used to disable
this behaviour.