Make rd.antievilmaid.dontforcestickremoval flag unnecessary
Automatically decide if the user should remove an AEM partition's
device, by applying the following criteria: Does the kernel consider
it removable (e.g. USB stick / SD card), *and* is it stored on a
different disk than /proc/cmdline's first rd.luks.uuid device?

lsblk is used to traverse the device hierarchy, and to get the
"removable" attribute. In case the kernel gets the latter wrong,
this can be overridden using the /anti-evil-maid/removable quirk
file in the AEM filesystem.