Use strict shell error checking
With so many changes, it's probably a good idea to switch
anti-evil-maid-unseal to 'set -e -o pipefail' to make all errors (even
in pipelines) cause an immediate failure.

dracut-lib.sh's getargs seems to be incompatible with strict error
checking, and interferes with 'set -x' debugging, too. Off with its
head!